Move s3 generator to cloud utils
