+++ blobfish/utils/cloud_utils.py
@@ -1,5 +1,7 @@
 import os
 import boto3
+from typing import Generator
+
 
 def clear_downloads(bucket: str, prefix: str):
     client = boto3.client(
@@ -27,3 +29,18 @@ def view_downloads(bucket: str, prefix: str):
         contents = page.get("Contents", [])
         for content in contents:
             print(client.get_object(Bucket=bucket, Key=content.get("Key")))
+
+
+def get_mirrored_content(bucket: str, prefix: str) -> Generator[dict, None, None]:
+    client = boto3.client(
+        service_name="s3",
+        aws_access_key_id=os.environ["AWS_ACCESS_KEY_ID"],
+        aws_secret_access_key=os.environ["AWS_SECRET_ACCESS_KEY"],
+        region_name=os.environ["AWS_DEFAULT_REGION"],
+    )
+    paginator = client.get_paginator("list_objects_v2")
+    for page in paginator.paginate(Bucket=bucket, Prefix=prefix):
+        contents = page.get("Contents", [])
+        for content in contents:
+            object = client.head_object(Bucket=bucket, Key=content.get("Key"))
+            yield object
